src/backend/requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.0
typing_extensions==4.11.0
debugpy==1.8.1
djangorestframework
django-cors-headers
google-cloud-speech
pydub
openai-whisper
asteroid-filterbanks>=0.4
einops>=0.6.0
huggingface_hub>=0.13.0
lightning>=2.0.1
omegaconf>=2.1,<3.0
pyannote.core>=5.0.0
pyannote.database>=5.0.1
pyannote.metrics>=3.2
pyannote.pipeline>=3.0.1
pytorch_metric_learning>=2.1.0
rich>=12.0.0
semver>=3.0.0
soundfile>=0.12.1
speechbrain>=1.0.0
tensorboardX>=2.6
torch>=2.0.0
torch_audiomentations>=0.11.0
torchaudio>=2.2.0
torchmetrics>=0.11.0
torchvision
pyannote.audio